requirements.txt
Django==5.0.7
django-model-utils==4.5.1
gunicorn==22.0.0
django-debug-toolbar==4.4.6
cryptography==44.0.2
Django==5.0.7
django-debug-toolbar==4.4.6
django-model-utils==4.5.1
l4py==0.1.9
royman-dotenv==1.1.2
